Move QueryClientProvider
#126
